refactor: 消除 orchestra/services/roles/runtime 扩展循环依赖
消除 4 个 L3 核心层循环依赖（issue #2125），使循环数从 8 个减少到 3 个：

### 目标循环（已消除）
- services → roles → services: scan_service.py 从 services/ 移至 roles/
- services → runtime → services: 拆分为两步修复：
  1. dag_service 跳过 if TYPE_CHECKING: 块（消除误报的 orchestra_status_service → runtime 边）
  2. orchestra_instance.py 从 runtime/ 移至 utils/（task_status_service 改从 utils 导入）

### 附加修复
- domain → runtime → domain: ServiceBase 从 runtime/service_protocol.py 移至
  domain/protocols/runtime_protocols.py，消除 orchestration_facade → runtime 依赖

### 变更文件
- analysis/dag_service.py: _extract_imports 跳过 TYPE_CHECKING 节点（正确建模运行时依赖）
- utils/orchestra_instance.py: 新建，包含 OrchestraInstanceInfo/read_instance_info/validate_instance
- runtime/orchestra_instance.py: 改为从 utils 重导出的存根
- roles/scan_service.py: 新建，从 services/scan_service.py 迁移而来
- services/scan_service.py: 删除（迁移至 roles/）
- domain/protocols/runtime_protocols.py: 新建，包含 ServiceBase
- runtime/service_protocol.py: 改为从 domain/protocols 重导出的存根
- domain/orchestration_facade.py: 改从 domain/protocols 导入 ServiceBase
- services/task_status_service.py: 改从 vibe3.utils.orchestra_instance 导入
- commands/scan.py, commands/internal.py: 更新 scan_service 导入路径
- tests: 更新 xfail 注释反映当前 3 个剩余循环
